move PollCtx to its own compilation unit
